+++ src/Websites/Websites/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated API version for WebApps operations from 2021-02-01 to 2026-03-15
+* Added Async Scaling support for App Service Plans in `New-AzAppServicePlan` and `Set-AzAppServicePlan`
+    - Added support for setting the `asyncScalingEnabled` Azure Resource Manager (ARM) property using `-AsyncScalingEnabled` (true or false)
 
 ## Version 4.0.0
 * Added ChangeSafety Support